fixed 'full_birds' and numbers to add postbreeding
there was some confusion about bringing in 'full_birds' as it was already culled. this is now fixed, and we cull it to passerines with good data coverage. 
Also the data simplification numbers are now fixed to account for the post-breeding weather, although it might be better to change them to column names as then I wouldn't have to fix the numbers when we make changes...